Latest features and bug fixes imported from stable

+++ includes/LinkCache.php
@@ -75,9 +75,11 @@ class LinkCache {
 		$this->clearLink( $title );
 	}
 	
-	function clearLink( $title ) {
-		global $wgMemc;
-		$wgMemc->delete( $this->getKey( $title ) );
+	function clearLink( $title ) 
+	{
+		global $wgMemc, $wgLinkCacheMemcached;
+		if( $wgLinkCacheMemcached )
+			$wgMemc->delete( $this->getKey( $title ) );
 	}
 
 	function suspend() { $this->mActive = false; }
@@ -114,8 +116,11 @@ class LinkCache {
 			wfProfileOut( $fname );
 			return 0; 
 		}
+		
+		$id = FALSE;
+		if( $wgLinkCacheMemcached )
+			$id = $wgMemc->get( $key = $this->getKey( $title ) );
 
-		$id = $wgMemc->get( $key = $this->getKey( $title ) );
 		if( $id === FALSE ) {
 			$sql = "SELECT cur_id FROM cur WHERE cur_namespace=" .
 			  "{$ns} AND cur_title='" . wfStrencode( $t ) . "'";
@@ -127,7 +132,8 @@ class LinkCache {
 				$s = wfFetchObject( $res );
 				$id = $s->cur_id;
 			}
-			$wgMemc->add( $key, $id, time()+3600 );
+			if( $wgLinkCacheMemcached )
+				$wgMemc->add( $key, $id, time()+3600 );
 		}
 		if ( 0 == $id ) { $this->addBadLink( $title ); }
 		else { $this->addGoodLink( $id, $title ); }
